+++ src/components/Button/components/ButtonWithTooltip.js
@@ -7,18 +7,16 @@ const ButtonWithTooltip = React.forwardRef(
   ({ tooltip, tooltipDisabled, disabled, ...props }, ref) => {
     const button = <Button ref={ref} disabled={disabled} {...props} />;
 
-    if (tooltip) {
-      return (
-        <Tooltip title={tooltip} translate>
-          {button}
-        </Tooltip>
-      );
-    }
-
     if (tooltipDisabled && disabled) {
       return (
         <Tooltip title={tooltipDisabled} translate>
           <span>{button}</span>
+        </Tooltip>
+      );
+    } else if (tooltip) {
+      return (
+        <Tooltip title={tooltip} translate>
+          {button}
         </Tooltip>
       );
     }