+++ src/components/Navigation/components/Item.js
@@ -2,5 +2,5 @@ import React from 'react';
 import Tab from '@material-ui/core/Tab';
 
 export default function Item(props) {
-  return <Tab {...props} />
+  return <Tab {...props} />;
 }